권한 위치 변경
